requirements.txt
REQUIREMENTS
1-python 3
3-pip
2-virtual environment python
3-Django 4
4-Pillow
5-Psycopg2
6-Psycopg2-binary
7-PostgeSQL 13

HOW TO RUN THIS PROJECT
1- ..\..\venv\Scripts\activat
2- python manage.py runserver